Added login signup database design
In this commit, we have added and solved several features and problems:

->Created a LogIn and Registration activity.
When the app is opened, the login page will be displayed by default. Users can login using their email and password (we will add a "forget password" option soon). If the user is a new user, they can click the "signup" button and be redirected to the signup page. In the signup page, users can signup using their name, email, and password, and their user profile will be created in our database. We have also added a background image to the login and registration pages.

->Created a "show uploaded image" button in the main activity.
When clicked, it will take the user to an "UserActivity" page.

->Created an UserActivity page to display the uploaded images by the user.
Here, users can see all the images they have uploaded till now with its upload time and geo-tags in a custom-list format (using recyclerview)(you must be logged in to open this page). We have taken reference from the following Github link for the recyclerview and custom list: https://github.com/foxandroid/RealtimeRecyclerView/tree/master/app/src/main/java/com/example/realtimerecyclerview and youtube link : https://youtu.be/M8sKwoVjqU0

->Stored and correctly fetched the download URL of the uploaded image.
In the previous version, the download URL of the image was not fetched correctly, so we have corrected that by using the getDownloadURL() method and adding an onSuccessListener on this method which generates the URI of the uploaded image. We have taken reference from the following StackOverflow link for the download URL: https://stackoverflow.com/a/37379251

->Created proper database storage in a nice and formatted way (multi-level nodes).
The structure of the storage is as follows:
users(root)->userid->{userId,username,email,password,joindate,uploadedImages->{datetime->{imageUrl,latitude,longitude,timeofupload}}}
We have taken reference from the following youtube video link for the database storage: https://youtu.be/M8sKwoVjqU0

Co-Authored-By: verma321 <92589137+verma321@users.noreply.github.com>
Co-Authored-By: Manish Kumar Sharma <105535356+shubhman20@users.noreply.github.com>
Co-Authored-By: UTKARSHANANDX <75621246+UTKARSHANANDX@users.noreply.github.com>